Added string utilities, and tested them.
- Utility to split a string with a given delimiter.
- Utility to strip a given char from a string.

Also, moved string_utils.hpp to ekat_string_utils.hpp, and
moved impl of some non-templated method to a cpp file.
